fix(hooks): Seperated responsability for hooks and clean imports
