+++ app/(client)/client/tasks/[id]/validate/page.tsx
@@ -4,11 +4,10 @@ import ReviewToast from '@/components/shared/review/ReviewToast';
 import { useTasksContext } from '@/components/shared/tasks/TaskProvider';
 import Button from '@/components/ui/Button/Button';
 import { useModalify } from '@/components/ui/Modal/hooks/useModalify';
-import { Modale } from '@/components/ui/Modal/Modale';
-import { Overlay } from '@/components/ui/Overlay/Overlay';
 import Spinner from '@/components/ui/Spinner/Spinner';
 import { useToasting } from '@/components/ui/Toast/useToasting';
-import { budgetCurrency, usePayment } from '@/hooks/useTasks';
+import { usePayment } from '@/hooks/usePayment';
+import { budgetCurrency } from '@/hooks/useTasks';
 import apiFetch from '@/lib/api';
 import { deliverablesWithVariedFiles, tasksA } from '@/lib/data';
 import { commissionPlateform, formatFrenchDateIntl, getInitials } from '@/lib/utils';
