+++ app/(prestataire)/prestataire/dashboard/page.tsx
@@ -5,7 +5,8 @@ import TaskCard from "@/components/dashboard/prestataire/TaskCard"
 import FlagTask from "@/components/shared/tasks/FlagTask"
 import Button from "@/components/ui/Button/Button"
 import Spinner from "@/components/ui/Spinner/Spinner"
-import { budgetCurrency, useDashboard } from "@/hooks/useTasks"
+import { useDashboard } from "@/hooks/useDashboard";
+import { budgetCurrency } from "@/hooks/useTasks"
 import { cn } from "@/lib/utils"
 import {  PDashboardData } from "@/types"
 import { ArrowRight } from "lucide-react"
