+++ hooks/useTasks.ts
@@ -1,10 +1,9 @@
 'use client'
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect } from 'react';
 import apiFetch from '@/lib/api';
-import type { ApplicationFormType, CDashboardData, Deliverable, PDashboardData, Review, ReviewProps, TaskFormType, TaskProps, ApplicationResponse, PrestataireSelectedData, Application } from '@/types';
+import type { Deliverable, Review, ReviewProps, TaskFormType, TaskProps } from '@/types';
 import { DeliveryFormProps } from '@/types/index';
 import { useRouter } from 'next/navigation';
-import { toast } from "react-hot-toast";
 import { useToasting } from '@/components/ui/Toast/useToasting';
 
 export interface UseTasksReturn<T = TaskProps> {
@@ -131,7 +130,6 @@ export function useTasks<T =  TaskProps>(id:string|undefined, skip:boolean=false
       return false;
   }
 
-
   useEffect(() => {
     if(skip) return;
 
@@ -148,206 +146,4 @@ export function useTasks<T =  TaskProps>(id:string|undefined, skip:boolean=false
     deleteTask,
     editTask  
   };
-}
-
-
-interface UseApplicationReturn {
-  application: ApplicationResponse[],
-  loading: boolean,
-  applyTotask: (applyData: ApplicationFormType) => Promise<void>,
-  rejectApplication: (application_id:string) => Promise<void>,
-  acceptApplication: (application_id:string) => Promise<void>,
-  getTaskApplication: (task_id:string, role:'client'|'prestataire') => Promise<void>,
-} 
-
-
-export function useApplication(): UseApplicationReturn {
-  const [application, setApplication] = useState<ApplicationResponse[]>([]);
-  const [loading, setLoading] = useState(true);
-  const {notify} = useToasting();
-  const router = useRouter();
-
-   const applyTotask = async (applyData: ApplicationFormType) => {
-        try{
-          setLoading(true);
-          const applyresponse = await apiFetch<ApplicationResponse>(`api/applications/apply`, applyData, 'POST');
-          
-          if(applyresponse.success){
-            notify('Votre candidature a été soumise avec succès.', 'success');
-            setApplication([applyresponse.data]);
-          }
-          else throw new Error(applyresponse.message) 
-        }catch(err){
-            notify(err instanceof Error ? err.message : 'Une erreur est survenue: Candidature non soumise!', 'error');
-        }finally{
-          setLoading(false);
-        }
-    }
-
-    const getTaskApplication = async (task_id:string, role:'client'|'prestataire'): Promise<void> => {
-      try{
-        setLoading(true);
-        const response = await apiFetch<ApplicationResponse[]>(`api/tasks/${task_id}/applications${role === 'prestataire' ? '/me':''}`);
-        if(response.success)
-          setApplication(response.data);
-        else
-          notify(response.message, 'error');
-      }catch(err){
-        notify(err instanceof Error ? err.message : 'Une erreur est survenue lors du chargement de vos candidatures!', 'error');
-      }finally{
-        setLoading(false);
-      }
-    }
-    const rejectApplication = async (application_id:string) => {
-      try{
-          setLoading(true);
-          const applyresponse = await apiFetch<null>(`api/applications/reject`, {id:application_id}, 'DELETE');
-          
-          if(applyresponse.success)
-            notify('Votre candidature a été retirée.', 'success');
-          else throw new Error(applyresponse.message) 
-        }catch(err){
-            notify(err instanceof Error ? err.message : 'Une erreur est survenue: Candidature non retirée!', 'error');
-        }finally{
-          setLoading(false);
-        }
-    }
-
-    const acceptApplication = async (application_id:string) => {
-      try{
-          setLoading(true);
-          const applyresponse = await apiFetch<null>(`api/applications/accept`, {id:application_id}, 'PUT');
-          
-          if(applyresponse.success)
-            notify('Votre candidature a été acceptée.', 'success');
-          else throw new Error(applyresponse.message) 
-        }catch(err){
-            notify(err instanceof Error ? err.message : 'Une erreur est survenue: Candidature non acceptée!', 'error');
-        }finally{
-          setLoading(false);
-        }
-    }
-
-
-    return {applyTotask, application, loading, rejectApplication, acceptApplication, getTaskApplication}
-}
-
-
-export interface UseDashboardReturn<T = CDashboardData | PDashboardData| undefined> {
-  loading: boolean
-  dashboardData: T | undefined
-  loadDashboard: () => void,
-  error: string | null
-}
-
-export function useDashboard<T = CDashboardData | PDashboardData | undefined>(
-  role: 'client' | 'prestataire' = 'client'
-): UseDashboardReturn<T> {
-  const [loading, setLoading] = useState(false);
-  const [dashboardData, setDashboardData] = useState<T | undefined>(undefined);
-  const [error, setError] = useState<string | null>(null);
-  const {notify} = useToasting();
-
-  const loadDashboard = useCallback(async () => {
-    if (loading) return;
-    try {
-      setLoading(true);
-      setError(null);
-      
-      const response = await apiFetch<T>(`api/dashboard/${role}`);
-      
-      if (response.success) {
-        setDashboardData(response.data);
-      } else {
-        throw new Error(response.message || 'Erreur de chargement');
-      }
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'Une erreur est survenue lors du chargement';
-      setError(message);
-      notify(message, 'error');
-    } finally {
-      setLoading(false);
-    }
-  }, [role, loading, notify]);
-
-  return {
-    loadDashboard,
-    loading,
-    dashboardData,
-    error,
-  };
-}
-
-export interface UsePaymentReturn {
-  loading: boolean,
-  proceedToPayment: (data: { application_id: string, task_id: string, prestataire_name: string }) => Promise<void>,
-  liberatefunds: (deliverable_id:string) => Promise<boolean>,
-  verifyPayment: (task_id:string)=>Promise<void>
-}
-
-export function usePayment<UsePaymentReturn >() {
-  const [loading, setLoading] = useState(true);
-  const router = useRouter();
-    const {notify} = useToasting();
-
-  const proceedToPayment = async (data: PrestataireSelectedData) => {
-  try{
-      setLoading(true);
-      const saveApplicant = await apiFetch<null>('/api/applications', data, 'POST');
-      if(saveApplicant.success){
-       notify('Candidature sauvegardée. Procédez au paiement.', 'success');
-        router.push(`/client/tasks/${data.task_id}/payment`);
-      }          
-      else throw new Error(saveApplicant.message)
-    }catch(err){
-        notify(err instanceof Error ? err.message : 'Une erreur est survenue lors de l\'acceptation de la candidature!', 'error');
-    }finally{
-      setLoading(false);
-    }
-  }
-
-  const liberatefunds = async (deliverable_id:string)=> {
-      try{
-      setLoading(true);
-      const liberate = await apiFetch<null>(`/api/deliverables/validate/${deliverable_id}`, undefined, 'POST');
-      if(liberate.success){
-        router.push(`/client/dashboard`);
-        notify('Votre mission est maintenant achevée', 'success'); 
-        return true
-      }          
-      else throw new Error(liberate.message)
-    }catch(err){
-        notify(err instanceof Error ? err.message : 'Une erreur est survenue lors de la liberation des fonds!', 'error');
-    }finally{
-      setLoading(false);
-    }
-    return false
-  }
-
-  const verifyPayment = async (task_id:string) => {
-      try{
-      setLoading(true);
-      const verify = await apiFetch<null>(`/api/tasks/${task_id}/payment/verify`, undefined, 'POST');
-      if(verify.success){
-        const deleteCookie = await apiFetch<null>('/api/applications', undefined, 'DELETE')  
-        if(deleteCookie.success)  
-          router.push('/client/dashboard');
-          notify('Paiement effectué avec succès.', 'success')
-          return true
-      }          
-      else throw new Error(verify.message)
-    }catch(err){
-        notify(err instanceof Error ? err.message : 'Erreur lors du paiement.', 'error');
-    }finally{
-      setLoading(false);
-    }
-    return false
-  }
-
-  return{
-    loading,
-    proceedToPayment,
-    liberatefunds,
-    verifyPayment
-  }
 }
